+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+## TBD
+
+### Enhancements
+
+* Exception messages will be truncated if they have a length greater than 3,072
+  | [#636](https://github.com/bugsnag/bugsnag-ruby/pull/636)
+  | [joshuapinter](https://github.com/joshuapinter)
+
 ## 6.18.0 (27 October 2020)
 
 ### Enhancements